Consolidate regions before unioning into accumulator
So.. We sometimes run into issues with `turf.union` and JSTS.. :(
Some of the results we get from country-coder are MultiPolygon regions containing
many countries.  It seems to help if we reduce the complexity of these regions
before we try to union them into the accumulated feature.
Try unioning region with itself to remove internal boundaries.