Let's give bun a try
Bun is a new JavaScript/TypeScript runtime with everything built in.
It seems like it could be a lot easier than using a bunch of tools.
It also is a bit faster and has some improvements to security.
This replaces node, c8, esbuild, taze, maybe some other things.

Bun has a builtin http server, so we can use it for local testing.
Just `bun start` or `bun run start` to test location-conflation on localhost.